Add customize emoji upload to embed
- customize gender, moves and teratype emoji can now be upload and display in embed
